+++ src/main/java/de/samply/app/ProjectManagerConst.java
@@ -1,0 +1,10 @@
+package de.samply.app;
+
+public class ProjectManagerConst {
+
+    public static final String APP_NAME = "Project Manager";
+
+    // REST Services
+    public final static String INFO = "/info";
+    public static final String ACTIONS = "/actions";
+}