+++ src/main/java/de/samply/db/model/ProjectUser.java
@@ -1,0 +1,33 @@
+package de.samply.db.model;
+
+
+import de.samply.user.ProjectRole;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "project_user", schema = "samply")
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ProjectUser {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.SEQUENCE)
+    @Column(name = "id", nullable = false)
+    private Long id;
+
+    @ManyToOne
+    @JoinColumn(name = "project_id", nullable = false)
+    private Project project;
+
+    @ManyToOne
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+
+    @Column(name = "project_role", nullable = false)
+    private ProjectRole projectRole;
+
+}