+++ src/main/java/de/samply/db/repository/ProjectBridgeheadRepository.java
@@ -1,0 +1,8 @@
+package de.samply.db.repository;
+
+import de.samply.db.model.ProjectBridgehead;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProjectBridgeheadRepository extends JpaRepository<ProjectBridgehead, Long> {
+
+}