+++ src/main/java/de/samply/db/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package de.samply.db.repository;
+
+import de.samply.db.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+}