+++ src/main/java/de/samply/frontend/Action.java
@@ -1,0 +1,9 @@
+package de.samply.frontend;
+
+public record Action(
+        Site site,
+        Module module,
+        String path
+) {
+
+}