+++ src/main/java/de/samply/frontend/Module.java
@@ -1,0 +1,4 @@
+package de.samply.frontend;
+
+public enum Module {
+}