+++ src/main/java/de/samply/frontend/Site.java
@@ -1,0 +1,8 @@
+package de.samply.frontend;
+
+public enum Site {
+    DASHBOARD,
+    FORM,
+    PROJECT,
+    CONSENT
+}