+++ src/main/java/de/samply/project/ProjectBridgeheadState.java
@@ -1,0 +1,9 @@
+package de.samply.project;
+
+public enum ProjectBridgeheadState {
+    CREATED,
+    ACCEPTED,
+    REJECTED
+
+    //TODO: Complete states
+}