+++ src/main/java/de/samply/project/ProjectConst.java
@@ -1,4 +1,0 @@
-package de.samply.project;
-
-public class ProjectConst {
-}