+++ src/main/java/de/samply/security/SecurityConfiguration.java
@@ -1,0 +1,34 @@
+package de.samply.security;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.annotation.Order;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.web.SecurityFilterChain;
+
+import static org.springframework.security.config.Customizer.withDefaults;
+import static org.springframework.security.oauth2.core.authorization.OAuth2AuthorizationManagers.hasScope;
+
+@Configuration
+@EnableWebSecurity
+@Order(1)
+public class SecurityConfiguration {
+
+    @Value("${app.security.enabled}")
+    private boolean isSecurityEnabled;
+
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+        return (isSecurityEnabled) ?
+                http.authorizeHttpRequests(authorize -> authorize
+                                .requestMatchers("/messages/**").access(hasScope("message:read"))
+                                .anyRequest().authenticated()
+                        )
+                        .oauth2Login(withDefaults())
+                        .build() :
+                http.authorizeRequests(authorize -> authorize.requestMatchers("/**").permitAll()).build();
+    }
+
+}