+++ src/main/java/de/samply/user/OrganisationRole.java
@@ -1,0 +1,8 @@
+package de.samply.user;
+
+public enum OrganisationRole {
+    RESEARCHER,
+    BRIDGEHEAD_ADMIN,
+    PROJECT_MANAGER_ADMIN,
+    REMOVED
+}