+++ src/main/java/de/samply/user/ProjectRole.java
@@ -1,0 +1,10 @@
+package de.samply.user;
+
+public enum ProjectRole {
+    CREATOR,
+    DEVELOPER,
+    PILOT,
+    FINAL,
+    BRIDGEHEAD_ADMIN,
+    PROJECT_MANAGER_ADMIN
+}